Get a value list as data frame from a study tree, list of trees, or studies object

Get a value list as data frame from a study tree, list of trees, or studies object

get_valueList_asDf_fromStudyTree(
  x,
  requiredFields = NULL,
  flattenVectorsInDf = TRUE,
  pathString_regex_select = NULL,
  pathString_regex_flatten = NULL,
  pathString_regex_explode = NULL,
  fieldname_regex_alwaysFlatten = NULL,
  silent = metabefor::opts$get("silent")
)

get_valueList_asDf(
  x,
  requiredFields = NULL,
  pathString_regex_select = NULL,
  flattenVectorsInDf = TRUE,
  pathString_regex_flatten = NULL,
  pathString_regex_explode = NULL,
  fieldname_regex_alwaysFlatten = NULL,
  silent = metabefor::opts$get("silent")
)

Arguments

x
The study tree, list of trees, or studies object

requiredFields
Fields that have to exist in the target entities (otherwise, the entity is excluded)

flattenVectorsInDf
The default action to apply for values not matching one of pathString_regex_flatten and pathString_regex_explode: to flatten by default, pass TRUE; to explode by default, pass FALSE.

pathString_regex_select
Regex that the target entities' path strings have to match (otherwise, the entity is excluded)

pathString_regex_flatten, pathString_regex_explode
Regular expressions matched against each entity node's path string (i.e. its path from the root, delimited by slashes). Vectors in entity nodes matching pathString_regex_flatten will be flattened into a single character string value; vectors in entity nodes matching pathString_regex_explode will be exploded into multiple rows.

fieldname_regex_alwaysFlatten
A regular expression to force flattening of fields regardless of matching to other regular expressions.

silent
Whether to be quiet or chatty.

Value

A dataframe

Examples